+++ backend/src/main/java/ar/edu/unpsjb/jbpe/model/dto/SitioDeRecoleccionDTO.java
@@ -40,7 +40,7 @@ public class SitioDeRecoleccionDTO {
     private BigDecimal altitud;
 
     @NotNull(message = "locacionId es obligatorio")
-    private Long locacionId;
+    private Integer locacionId;
 
     private String nombreLocacion;
